requirements.txt
# Flask and extensions
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
python-dotenv==0.21.0

# MongoDB
pymongo==4.2.0
dnspython==2.2.1

# Security
bcrypt==3.2.2
PyJWT==2.6.0

# Development tools
pytest==7.1.3
pytest-flask==1.2.0
pytest-cov==4.0.0

# Utilities
Werkzeug==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
cffi==1.15.1
pycparser==2.21
